Fix tree node names and depth limiting
Fix up the way in which tree traversal depth is calculated (although
probably more needs to be done here).

Remove a todo relating to graph image uploading that is now handled by
the GraphViz extension.

Add some comments to the generated GraphViz output for debugging.

Fix the names of nodes that are used to link children and parents,
so that they work where there is only one parent or other duplicate
names.

Add tests.

Change-Id: I2c8031901ff1cdb39e4ea2bcd602dec8df3ec7dd