Simplify tree links to non-existing people
Links in GraphViz graphs to people who don't exist are done as
external links (with single brackets). They were including link
text as well, but this serves no purpose because the displayed
text is provided by the separate 'label' parameter. This change
removes the link text and leaves the single brackets.

Change-Id: Iaec33a3697574244e9ced2deccbdb68e007ccda2