Make tree names/IDs include depth
Tree names are used as image map IDs, and so must be unique within
a page. They were using the concatenation of names to do this,
but in the (rare) case that the same set of ancestors and
descendants is used on one page, but with different depths, there
would be a clash. This resolves this by extending the concatenation
to include the two depth parameters.

Change-Id: I4a180a8cb108423b38b408c5b746cda0dff9b5d0